+++ actions/google-drive/CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Added
 
-- Action `get_file_by_id` attaches file into the Files
+- Actions `get_file_by_id` and `get_file_contents` attach a file into the Files.
+  - Custom filename can be set
+  - Attach can be disabled
 
 ## [1.2.1] - 2025-06-18